refactor: rename utils class file
