+++ src/utils/index.ts
@@ -1,3 +1,3 @@
-import Utils from "./Utility";
+import Utils from "./text_utils";
 export { Utils };
 export { isNumber, convertToNumber } from "./number";
