+++ src/utils/text_utils.ts
